Fix intermittent failing test for ./cmd/goatcounter (hopefully)
I've never seen it happen locally, but this test fails regularly on
Travis https://travis-ci.org/zgoat/goatcounter/builds/638066722

	$ go test -race -coverprofile=coverage.txt -coverpkg=./... ./...
	ok  	zgo.at/goatcounter	3.622s	coverage: 21.9% of statements in ./...
	?   	zgo.at/goatcounter/acme	[no test files]
	?   	zgo.at/goatcounter/cfg	[no test files]
	Goatcounter version dev
	17:47:57 INFO: -stripe not given or doesn't contain all keys; billing disabled
	17:47:57 INFO: restarting "/tmp/go-build352383128/b183/goatcounter.test" when it changes (additional dirs: ./tpl)
	17:47:57 zdb: INFO: database "/tmp/goatcounter125149924/goatcounter.sqlite3" doesn't exist; loading new schema
	FAIL	zgo.at/goatcounter/cmd/goatcounter	5.141s
	?   	zgo.at/goatcounter/cmd/goatcounter-reindex	[no test files]
	ok  	zgo.at/goatcounter/cron	3.271s	coverage: 18.2% of statements in ./...
	ok  	zgo.at/goatcounter/handlers	3.529s	coverage: 42.7% of statements in ./...
	?   	zgo.at/goatcounter/pack	[no test files]
	FAIL
	The command "go test -race -coverprofile=coverage.txt -coverpkg=./... ./..." exited with 1.

I suspect this is because of the time.Sleep() (there is no "listening on
..." in the output). Travis performance can be inconsistent.

Hopefully this will fix it; time will tell.

+++ cmd/goatcounter/main_test.go
@@ -5,12 +5,16 @@
 package main
 
 import (
+	"bufio"
 	"flag"
 	"io/ioutil"
 	"os"
+	"strings"
 	"syscall"
 	"testing"
 	"time"
+
+	"zgo.at/ztest"
 )
 
 func TestMain(t *testing.T) {
@@ -33,12 +37,20 @@ func TestMain(t *testing.T) {
 	// Reset flags in case of -count 2
 	flag.CommandLine = flag.NewFlagSet(os.Args[0], flag.ExitOnError)
 	os.Args = []string{"goatcounter",
+		"-prod", "-smtp", "dummy",
 		"-dbconnect", tmpdb,
 		"-listen", "localhost:31874"}
 
+	out, reset := ztest.ReplaceStdStreams()
+	defer reset()
 	go func() {
-		time.Sleep(3000 * time.Millisecond)
-		syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+		scanner := bufio.NewScanner(out)
+		for scanner.Scan() {
+			if strings.Contains(scanner.Text(), "listening on") {
+				time.Sleep(100 * time.Millisecond)
+				syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+			}
+		}
 	}()
 
 	main()
